requirements.txt
openai
streamlit
fastapi
uvicorn
chromadb
sentence-transformers
PyPDF2
requests
beautifulsoup4
docx
pptx
pandas
numpy
scikit-learn
python-dotenv
git+https://github.com/openai/whisper.git

